pyproject.toml
[project]
name = "reelioo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
]
